Add PipelineBase class
