+++ OnnxStack.Core/Config/OnnxModelConfig.cs
@@ -5,7 +5,6 @@ namespace OnnxStack.Core.Config
 {
     public record OnnxModelConfig
     {
-        public OnnxModelType Type { get; set; }
         public string OnnxModelPath { get; set; }
 
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
