+++ OnnxStack.Core/Config/OnnxModelSetConfig.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.ML.OnnxRuntime;
-using System.Collections.Generic;
 
 namespace OnnxStack.Core.Config
 {
@@ -13,6 +12,5 @@ namespace OnnxStack.Core.Config
         public int IntraOpNumThreads { get; set; }
         public ExecutionMode ExecutionMode { get; set; }
         public ExecutionProvider ExecutionProvider { get; set; }
-        public List<OnnxModelConfig> ModelConfigurations { get; set; }
     }
 }
